requirements.txt
apache-airflow[pandas,postgres]>=2.5.0
astro-sdk-python[postgres]>=1.6.0
openpyxl>=3.0.0